requirements.txt
browser_cookie3==0.18.0
python-dotenv==1.0.0
requests==2.28.2
spotipy==2.19.0
